# [alpaka]  Use alpaka memory fences 