This small PR adds the new alpaka memory fence functionality (introduced in [alpaka v0.8.0](https://github.com/alpaka-group/alpaka/blob/develop/docs/source/basic/library.rst#memory-fences)) to AlpakaCore. Validated with serial and cuda backends, the throughput remains unchanged.


<img width="746" alt="alpaka_mem_fence" src="https://user-images.githubusercontent.com/41752419/155764783-2fdb9f60-6af2-4ebb-b275-c4f13d9b6e88.png">