@fwyzard 2022-02-25
@tonydp03 rather than simplifying the definition of `cms::alpakatools::threadfence`, can you replace it everyehwre with the corresponsing call to `alpaka::mem_fence` ?

Also, can you comment on the choice of `alpaka::memory_scope::Device{}` ?