@tonydp03 2022-02-28
Yes, I was not sure if it was better to delete the `threadfence` library and just use `alpaka::mem_fence` directly where needed.
About the choice, memory fences in Alpaka can be issued on the block or device level through `alpaka::memory_scope::Block` and `alpaka::memory_scope::Device`. In pixeltrack standalone, memory fences were applied by means of ``std::atomic_thread_fence` for TBB (alpaka::mem_fence specialization for CPU doesn't change on the `memory_scope selected`) and `__threadfence` for CUDA (which is issued on the `device` level according to CUDA [documentation](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#memory-fence-functions)).