@fwyzard 2022-03-04
@tonydp03 thanks for the latest update.

I've repeated the measurements on a Tesla T4 with an AMD EPYC 75F3 32-core processor.

For `alpaka --cuda` I didn't see any significant difference, the throughput is always within ~0.1% -- which is expected, since both versions end up calling `__threadfence`.

For `alpaka --serial` I start seeing a measurable difference, of the order of 2% with 12 threads and 5% with 20 threads.
Comparing the code, the current implementation of `cms::alpakatools::threadfence()` evaluates to nothing in the cpu serial case, while `alpaka::mem_fence(acc, alpaka::memory_scope::Device{})` does call `std::atomic_thread_fence(std::memory_order_acq_rel)`.

I suspect that we need an additional case to `memory_scope::Block` and `memory_scope::Device`, something like `memory_scope::Grid`:
  - call `__threadfence` for the CUDA/HIP backend, because they don't provide a grid-wise atomic
  - call `std::atomic_thread_fence(...)` for the TBB and other parallel CPU backends
  - call nothing for the serial CPU backend, because the execution is fully serial at the block and grid level

Could you discuss the idea with the alpaka developers ?