@tonydp03 2022-03-04
I see, on the block level `alpaka::mem_fence` does nothing for Cpu serial, but on device level it was conceived to call `std::atomic_thread_fence` in case of synchronization with other serial kernels. I think that the addition of `memory_scope::Grid` makes sense. I'll discuss with the other developers along with the naming scheme.