@fwyzard 2022-03-11
And here is the new comparison:
	
threads | events | alpaka --serial (master) | alpaka --serial (grid-wise) | relative throughput
--: | --: | --: | --: | --:
2 | 20000 | 80.9 ev/s ± 0.5 ev/s | 80.6 ev/s ± 0.1 ev/s | 99.5% ± 0.6%
4 | 20000 | 150.5 ev/s ± 0.6 ev/s | 149.5 ev/s ± 0.4 ev/s | 99.3% ± 0.5%
6 | 20000 | 228.5 ev/s ± 4.5 ev/s | 230.7 ev/s ± 0.8 ev/s | 100.9% ± 2.0%
8 | 20000 | 301.0 ev/s ± 1.1 ev/s | 301.5 ev/s ± 0.8 ev/s | 100.1% ± 0.5%
10 | 20000 | 372.9 ev/s ± 10.0 ev/s | 373.1 ev/s ± 6.5 ev/s | 100.1% ± 3.2%
12 | 20000 | 446.5 ev/s ± 1.1 ev/s | 446.1 ev/s ± 1.1 ev/s | 99.9% ± 0.4%
14 | 20000 | 525.2 ev/s ± 2.5 ev/s | 520.0 ev/s ± 7.8 ev/s | 99.0% ± 1.6%
16 | 20000 | 594.1 ev/s ± 2.1 ev/s | 594.7 ev/s ± 0.5 ev/s | 100.1% ± 0.4%
18 | 20000 | 671.6 ev/s ± 1.6 ev/s | 659.3 ev/s ± 6.1 ev/s | 98.2% ± 0.9%
20 | 20000 | 740.8 ev/s ± 1.7 ev/s | 739.4 ev/s ± 2.2 ev/s | 99.8% ± 0.4%