@tonydp03 2022-03-14
Perfect. Now the performance of `alpaka --serial` seems the same as for `serial` within acceptable %.